separate app entrypoint from main
