+++ test/widget_test.dart
@@ -7,8 +7,7 @@
 
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-
-import 'package:pedax/main.dart';
+import 'package:pedax/app.dart';
 
 void main() {
   testWidgets('Counter increments smoke test', (tester) async {
